allow cli and json output
