+++ src/config/mod.rs
@@ -17,7 +17,7 @@ mod workload;
 
 pub use client::Client;
 pub use debug::Debug;
-pub use general::{General, OutputFormat};
+pub use general::General;
 pub use protocol::Protocol;
 pub use pubsub::Pubsub;
 pub use target::Target;
